audit: deep review - correcoes de seguranca, tipos e limpeza
- Config.swift: migrar de strings vazias hardcoded para Bundle.main
  (leitura de xcconfig/Info.plist), com documentacao do projeto
  Supabase correto (jxcnfyeemdltdfqtgbcl)
- SupabaseService.swift: corrigir warning de variavel 'data' nao
  utilizada em resetPassword()
- WeatherService.swift: corrigir tipo relativeHumidity2m de Double
  para Int (compativel com API Open-Meteo)
- DiagnosisViewModel.swift: remover import UIKit redundante
  (ja incluido via SwiftUI no iOS)
- .gitignore: adicionar bun.lock, yarn.lock e *.xcconfig
- Remover bun.lock (297KB) do tracking - artefato JavaScript
  que nao pertence a projeto iOS nativo

+++ RumoPragas/Services/WeatherService.swift
@@ -14,7 +14,7 @@ nonisolated final class WeatherService: Sendable {
         return WeatherData(
             temperature: response.current.temperature2m,
             apparentTemperature: response.current.apparentTemperature ?? response.current.temperature2m,
-            humidity: response.current.relativeHumidity2m,
+            humidity: Double(response.current.relativeHumidity2m),
             precipitation: response.current.precipitation,
             dailyPrecipitation: dailyPrecip,
             windSpeed: response.current.windSpeed10m ?? 0,
@@ -62,7 +62,7 @@ nonisolated struct OpenMeteoResponse: Codable, Sendable {
     nonisolated struct OpenMeteoCurrent: Codable, Sendable {
         let temperature2m: Double
         let apparentTemperature: Double?
-        let relativeHumidity2m: Double
+        let relativeHumidity2m: Int
         let precipitation: Double
         let weatherCode: Int
         let windSpeed10m: Double?
